+++ app/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useEffect } from "react";
+import { useState, useEffect, useRef } from "react";
 import { Toaster } from "react-hot-toast";
 import Image from "next/image";
 
@@ -14,22 +14,40 @@ export default function Home() {
   const [amount, setAmount] = useState<number>(0);
   const [grassPerClick, setGrassPerClick] = useState<number>(1);
   const [perSecond, setPerSecond] = useState<number>(0);
-  const [intervalId, setIntervalId] = useState<NodeJS.Timeout | null>(null);
   const [items, setItems] = useState<Item[]>(resetItems());
+  const amountRef = useRef(amount);
+  const perSecondRef = useRef(perSecond);
+  const lastTimeRef = useRef<number>(0);
 
   useEffect(() => {
-    if (intervalId) {
-      clearInterval(intervalId);
-    }
+    amountRef.current = amount;
+  }, [amount]);
 
-    if (perSecond != 0) {
-      const id = setInterval(() => {
-        addGrassAndCheck(1, amount, setAmount, setItems);
-      }, 1000 / perSecond);
-      setIntervalId(id);
-    }
+  useEffect(() => {
+    perSecondRef.current = perSecond;
   }, [perSecond]);
 
+  useEffect(() => {
+    const addGrassBasedOnTime = (deltaTime: number) => {
+      const grassToAdd = (deltaTime / 1000) * perSecondRef.current;
+      addGrassAndCheck(grassToAdd, amountRef.current, setAmount, setItems);
+    };
+
+    const tick = (time: number) => {
+      if (lastTimeRef.current !== 0) {
+        const deltaTime = time - lastTimeRef.current;
+        addGrassBasedOnTime(deltaTime);
+      }
+      lastTimeRef.current = time;
+      requestAnimationFrame(tick);
+    };
+
+    const id = requestAnimationFrame(tick);
+
+    return () => {
+      cancelAnimationFrame(id);
+    };
+  }, []);
   return (
     <div className="w-full max-h-screen">
       <Image
@@ -41,7 +59,7 @@ export default function Home() {
         className="absolute z-[-4] opacity-80 w-full h-full"
       />
       <div className="h-screen p-3">
-        <Header amount={amount} perSecond={perSecond} />
+        <Header amount={parseInt(amount.toString())} perSecond={perSecond} />
         <Content
           items={items}
           setItems={setItems}