+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # grass-clicker-next changelog
 
+## 0.0.2 (16/05/24)
+
+- added more items
+- item buying confirmation
+- fixed bug where you would not get the full amount of grass per second if it were too high
+- updated changelog
+
 ## 0.0.1 (14/05/24)
+
 - switched to use bun instead of npm
 - used better and more consistent code style
 - finished the basic functionality of the game
+- made a changelog